+++ jest.setup.js
@@ -41,3 +41,5 @@ jest.mock('react-native-safe-area-context', () => {
     initialWindowMetrics: { insets, frame },
   };
 });
+
+